refactor: migrate reorder mutation to transaction
